Module probables::tree

Tree is common data structure that are used to represent data … i.e file or directory Tree module is not generic , currently it will allow string type data. But soon we will translate into generics

Structs

CachedResult

CacheResults return data quickly. CacheResults you don’t need to compute everytime

PrimaryNode

PrimaryNode (Branch) is a connector which connect leaves. Branch hold leaves and ants move easily. Currently backtracking is not implemented

Enums

Branch

Branch “Enums” have three states either { “Parent” => Leave attached with tip of branch “Left” => Leave attached with left side of branch “Right” => Leave attached with right side of branch }

Results

Results “Enums” either Ok , Err Ok => Data validation Err => Data validation failure

Statics

DEPTH

HEIGHT

Traits

Tree