Trait probables::tree::Tree

pub trait Tree {
    fn branch_leaf(
        &mut self, 
        node: PrimaryNode, 
        branch: Branch
    ) -> Vec<PrimaryNode>;

    fn new_branch(
        &mut self, 
        node: PrimaryNode, 
        branch: Branch, 
        iter: u8
    ) -> Vec<PrimaryNode>;

    fn depth(&mut self, branch: Branch) -> usize;

    fn recursive_leaf(&mut self, branch: Branch, index: usize) -> usize;

    fn eagleview(
        &mut self, 
        branch: Branch, 
        data: String, 
        index: usize
    ) -> Results;
}

Required methods

fn branch_leaf(&mut self, node: PrimaryNode, branch: Branch) -> Vec<PrimaryNode>

Branch Leaf is a receiver provided by tree module & tree trait. Branch Leaf take many parameters as arguments { Muatble self object, Node Object and branch enumerate}

fn new_branch(
&mut self,
node: PrimaryNode,
branch: Branch,
iter: u8
) -> Vec<PrimaryNode>

New Branch take mutable object , Node Object , Branch Enumerate and branch index. New Branch return Vector of Node Object.
Branch Enumerate State decide where will new branch emeraged. In Parent case empty branch object will return.

fn depth(&mut self, branch: Branch) -> usize

Depth is a receiver method which takes in mutable self Object and Branch enumerate. Depth return usize value. Depth return Bob family children and grandchildren.

fn recursive_leaf(&mut self, branch: Branch, index: usize) -> usize

Recursive Receiver takes a bunch of arguments as parameters Muatble Self Object, Branch Enumerate and index value. Recursive Receiver return value. Recursive Receiver iterates leaf in other words ant walk over leaves.

fn eagleview(&mut self, branch: Branch, data: String, index: usize) -> Results

Eagleview Reciver takes a bunch of arguments as parameters. Muatble Self Object , Branch Enumerate, Data in string and index as value. EagleView Receiver return Result Enumerate. Data should not be empty.