Module probables::tree

Tree is a module which contain collection of objects, functions and traits etc. Tree have static varaibles {Height and Depth}. More Learn about them discussed in static section below

Structs

CachedResult

Cache Result is an generic object. This object contains in memory object and state which is in String. Cache implement clousre. Check the docs if you’re curiois in. This object implement Fn trait.

PrimaryNode

Primary Node is an object. Few deriveable traits attached with primary node object. User should implement these traits. Primary Node have three fields. Data is the first field. Primary Node have two node object {Left or Right} These nodes are dynamic in nature.

Enums

Branch

Branch Enumeration have three Variants. Branch Enumeration is used when programmer want to play around with tree structure. Like Result it’s also derivable with tarit called Debug. Programer must enforces these traits properties. Enumerate State are Left, Right & Parent.

Results

Results have only one derivable traits Debug. Results Enumeration are used during executation of a program. OK and Error are two enumerate state in any program. When a program fail during executation then program in Error state, otherwise program is in other state.

Statics

DEPTH

Another static variable is Depth. Depth is level pointer. Means If a node name with (Bob) and alice asked how much members in your family. Bob will answer like I have 1 son and a daughter and both have 2 sons and 2 daughters. My family size will be 6.

HEIGHT

Height is a static variable. Height maximum space for Bob Family. Bob Family have 6 children, 1 wife 1 himself. Height of Bob Family or Family space will be 8. In Height parents are also include while depth parents are excluded.

Traits

Tree